+++ lib/presentation/viewmodels/home/detail_viewmodel.dart
