+++ lib/presentation/views/home/detail_screen.dart
